code split forms, register and login functionality complete
